break up project_settings file
